Move `modules` into independent folders
