+++ app.py
@@ -1,16 +1,19 @@
-import requests
 from flask import Flask, render_template, request, jsonify
 
-from modules.env import PORT, MONGO_URL, DATABASE_NAME, COLLECTION_USER, COLLECTION_CAPSULE, COLLECTION_SELECT, TOKEN_SECRET, TOKEN_ALGORITHM
-from modules.database import getMongoClient
-from modules.form import getSuccessForm, getFailureForm
+# Provider : 특정 기능을 공급하는 함수들
+from src.modules.provider.mongo_provider import getMongoClient
+from src.modules.provider.form_provider import getSuccessForm, getFailureForm 
+from src.modules.config.config_provider import PORT, MONGO_URL, DATABASE_NAME, COLLECTION_USER, COLLECTION_CAPSULE, COLLECTION_SELECT, TOKEN_SECRET, TOKEN_ALGORITHM
 
-from modules.tokenizer import getToken
-from modules.validate import validate_name, validate_password
+# Validator : 특정 값의 유효성을 테스트하는 함수들
+from src.modules.vadliator.form_validator import validate_name, validate_password
+
+# Tokenizer : 토큰을 생성하는 함수
+from src.modules.auth.tokenizer import getToken
+
 
 app = Flask(__name__)
 database = getMongoClient(MONGO_URL)[DATABASE_NAME]
-
 
 #결과창
 @app.route('/result', methods=["GET"])
